Faster pelf
Faster than normal pelf because it does not use bsae64, instead it cats the tar file directly to the blob, downside is that editing this file after the fact to change the start options of the blob, or the logic such as, where to decompress the libraries to, cannot be done because most text editors mess up binary data. Still, this is the faster, and lightweight version of pelf, since it generates blobs that weight almost the same as the binary + its libraries. Its worth remembering that pelf provides two options which can be set with env variables, please read the scripts.